git raw add: write a log entry for each added path
Each add command will now write a tab-separated log entry into a file
called "log" in the store directory of the added path. The format of
each log entry is currently:

<date and time>[tab]add[tab]<repo-name>[tab]<path>

where <repo-name> is the name of the directory hosting the repository,
i.e., basename of the git repository's topdir.