Major internal redesign
Since we want to be able to use multiple content stores, it is not
enough to store links to the content stores under
.git/git-raw/store. Such a design would imply that links in the
working directory would have to change for each user depending on
his/her local stores.

Instead, we will use a local directory called .git/git-raw/index that
will contain links to individual SHA1 directories in the local
stores. This way, all links stored in git will remain unchanged when
changing store configuration.

A file with SHA1 abc..123 will be stored in a content store as:

<store>/ab/abc..123/raw

The index directory will contain a direct link to the SHA1-specific
directory in the content store:

.git/git-raw/index/ab/abc...123 --> <store-path>/ab/abc...123/

The Git repository will then manage a link to

<rel-path-to-top>/.git/git-raw/index/ab/abc...123/raw

The target of the link will then be the same in all clones of the
repostory, but requiring some kind of fixup command that will populate
the index/ directory with links to content store SHA1 directories.

changes in this commit:

- rewrote major portions to accomodate redesign as described
  above. Major affected parts include:
  - the Git class
  - add command
  - add-store command
- get_config() renamed to get_git_config()
- configparser module is used to manage .git/git-raw/config
- store_mkname() returns only the full path to the raw file
- new function check_storename() for ensuring valid store names
- improve Git class documentation
- Git class constructor does more checking of the internal git-raw state
- fixes path_is_in_repo()