os.rename to shutil.move
Using shutil.move() instead of os.rename(). If src and dest are
on different filesystems os.rename will fail. shutil.move
uses os.rename if the files are on the same filesystem, otherwise
it copies src to dst and removes dst.

example: if the data store is on one nfs share and the git repo
on another using os.rename won't work.